fixed button to add notes, started fixing edit notes
