+++ src/components/atoms/DummyData.ts
@@ -1,0 +1,1 @@
+export const initialData = [{ noteId: 0, text: "" }];
